predicted rating is float
git-svn-id: http://subversion.assembla.com/svn/olimpo@233 080cfd25-d5d9-4c59-82ea-70d02a200d8b

+++ trunk/prototipo/experimento/db/migrate/20091125043754_create_system_recommendations.rb
@@ -4,7 +4,7 @@ class CreateSystemRecommendations < ActiveRecord::Migration
       t.integer :user_id
       t.string :algorithm
       t.integer :product_id
-      t.integer :predicted_rating
+      t.float :predicted_rating
 
       t.timestamps
     end
